+++ include/endstone/endstone.hpp
@@ -105,6 +105,7 @@ static_assert(_ITERATOR_DEBUG_LEVEL == 0,
 #include "event/player/player_drop_item_event.h"
 #include "event/player/player_emote_event.h"
 #include "event/player/player_event.h"
+#include "event/player/player_exp_change_event.h"
 #include "event/player/player_game_mode_change_event.h"
 #include "event/player/player_interact_actor_event.h"
 #include "event/player/player_interact_event.h"